Moved APError to seperate file
