+++ SwiftAudio/Classes/AVPlayerWrapper/AVPlayerWrapper.swift
@@ -21,24 +21,6 @@ protocol AVPlayerWrapperDelegate: class {
     
 }
 
-public struct APError {
-    
-    enum LoadError: Error {
-        case invalidSourceUrl(String)
-    }
-    
-    enum PlaybackError: Error {
-        case noLoadedItem
-    }
-    
-    enum QueueError: Error {
-        case noPreviousItem
-        case noNextItem
-        case invalidIndex(index: Int, message: String)
-    }
-    
-}
-
 class AVPlayerWrapper {
     
     struct Constants {
